==> code/0.eda.ipynb <==
import requests
import bs4 as bs
import pandas as pd

# url = 'https://thuvienphapluat.vn/page/tim-van-ban.aspx?keyword=&area=0&type={}&status=0&lan=1&org=0&signer=0&match=True&sort=1'
# cate=[
#  ('46', 'Báo cáo'),
#  ('1', 'Chỉ thị'),
#  ('26', 'Công điện'),
#  ('31', 'Điều ước quốc tế'),
#  ('6', 'Hiến pháp'),
#  ('9', 'Hướng dẫn'),
#  ('45', 'Kế hoạch'),
#  ('34', 'Lệnh'),
#  ('10', 'Luật'),
#  ('11', 'Nghị định'),
#  ('13', 'Nghị quyết'),
#  ('14', 'Pháp lệnh'),
#  ('17', 'Quyết định'),
#  ('18', 'Sắc lệnh'),
#  ('21', 'Thông báo'),
#  ('23', 'Thông tư'),
#  ('24', 'Thông tư liên tịch'),
#  ('40', 'Văn bản hợp nhất'),
#  ('33', 'Văn bản khác'),
#  ('32', 'Văn bản WTO')]

# results = []

# for cat in cate:
#     res = requests.get(url.format(cat[0]))
#     soup = bs.BeautifulSoup(res.text)
#     total = soup.select('#lbTotal')
#     total = int(total[0].text)
#     print(cat[1], total)
#     results.append((cat[1], total))

# df = pd.DataFrame(results, columns=['type', 'total'])
# #sort 
# df = df.sort_values(by=['total'], ascending=False)
# df.to_csv('data/tvpl_loaivanban.csv', index=False,header=False)

# df.plot(
#     kind='pie',
#     y='total',
#     labels=df['type'],
# )

# url = "https://thuvienphapluat.vn/page/tim-van-ban.aspx?keyword=&area=0"
# res = requests.get(url)
# soup=bs.BeautifulSoup(res.text)
# results = []
# li=soup.select_one('#tree').select('li')
# for l in li:
#     a = l.select('a')
#     if len(a) > 1:
#         print(a[0].text, a[1].text)
#         results.append((a[0].text, int(a[1].text)))
# df = pd.DataFrame(results, columns=['linhvuc', 'total'])
# df = df.sort_values(by=['total'], ascending=False)
# df.to_csv('data/tvpl_linhvuc.csv', index=False,     header=False)

import plotly.express as px

# plot 
df = pd.read_json('./answers.jsonl', lines=True)
df.head()

len(df)

import re

random = df.sample(1)
# take 2 col cauhoi, traloi
random = random[['cauhoi', 'traloi']]
random = random.reset_index(drop=True)
# remove duplicate newline in both col
random['cauhoi'] = random['cauhoi'].apply(lambda x: re.sub(r'\n+', '\n', x))
random['cauhoi'] = random['cauhoi'].apply(lambda x: re.sub(r'\n+', '\n', x))
random['traloi'] = random['traloi'].apply(lambda x: re.sub(r'\n+', '\n', x))
random['traloi'] = random['traloi'].apply(lambda x: re.sub(r'\n+', '\n', x))
# strip
random['cauhoi'] = random['cauhoi'].apply(lambda x: x.strip())
random['traloi'] = random['traloi'].apply(lambda x: x.strip())

print(random['cauhoi'].values[0])
print(random['traloi'].values[0])

# plot piechart on df linhvuc
# show value count
fig = px.pie( values=df['linhvuc'].value_counts().values, names=df['linhvuc'].value_counts().index,  width=700, height=400,
                )
fig.update_traces(textposition='inside', textinfo='percent')

# to image high quality
fig.write_image("lcauhoitheoinhvuc.svg",)
fig.show()
