==> code/1.process.ipynb <==
import pandas as pd
import json
import re
import json
from bs4 import BeautifulSoup
import bs4
import requests
import os
from tqdm import tqdm
import pickle
from law_query_private import LawQuery,Tree
from slugify import slugify
import gzip

# get luat hien hanh
req = requests.get('https://thuvienphapluat.vn/chinh-sach-phap-luat-moi/vn/thoi-su-phap-luat/tu-van-phap-luat/42248/danh-muc-luat-bo-luat-hien-hanh-tai-viet-nam')
soup = BeautifulSoup(req.text, 'html.parser')
luat_hien_hanh = soup.select('.newcontent a')
urls = []
for i in luat_hien_hanh:
    urls.append(i['href'])
urls = urls[1:]
# save to urls.txt
with open('urls.txt', 'w') as f:
    f.write('\n'.join(urls))

urls = []
with open('urls2.txt', 'r') as f:
    urls = f.read().split('\n')

def get_text(url):
    # get metadata
    metadata_url='https://thuvienphapluat.vn/AjaxLoadData/LoadLuocDo.aspx?LawID={}&IstraiNghiem=True'
    id = url.replace('.aspx','').split('-')[-1]
    req=requests.get(metadata_url.format(id),headers={'referer': url})
    if req.status_code != 200:
        raise Exception('Error when get metadata')
    soup = BeautifulSoup(req.text, 'html.parser')
    metadata = soup.select_one('#viewingDocument')
    metadata = {
        'ten_van_ban': metadata.select_one('#viewingDocument > div:nth-child(1)').getText(strip=True), 
        'so_hieu_van_ban': metadata.select_one('#viewingDocument > div:nth-child(2) > div.ds.fl').getText(strip=True),
        'loai_van_ban': metadata.select_one('#viewingDocument > div:nth-child(3) > div.ds.fl').getText(strip=True),
        'linhvuc': metadata.select_one('#viewingDocument > div:nth-child(4) > div.ds.fl').getText(strip=True),
        'noi_ban_hanh': metadata.select_one('#viewingDocument > div:nth-child(5) > div.ds.fl').getText(strip=True),
        'nguoi_ky': metadata.select_one('#viewingDocument > div:nth-child(6) > div.ds.fl').getText(strip=True),
        'ngay_ban_hanh': metadata.select_one('#viewingDocument > div:nth-child(7) > div.ds.fl').getText(strip=True),
        'ngay_hieu_luc': metadata.select_one('#viewingDocument > div:nth-child(8) > div.ds.fl').getText(strip=True),
        'ngay_cong_bao': metadata.select_one('#viewingDocument > div:nth-child(9) > div.ds.fl').getText(strip=True),
        'so_cong_bao': metadata.select_one('#viewingDocument > div:nth-child(10) > div.ds.fl').getText(strip=True),
        'tinh_trang': metadata.select_one('#viewingDocument > div:nth-child(11) > div.ds.fl').getText(strip=True),
    }
    req=requests.get(url)
    soup = BeautifulSoup(req.text, 'html.parser')
    html = soup.find('div', {'class':'content1'})
    for element in html(text=lambda text: isinstance(text, bs4.Comment)):
        element.extract()
    for a in html.find_all('a', href=lambda x: x and x[0] == '#'):
        a.extract()
    for s in html.find_all(['script','style']):
        s.extract()
    ps = html.find_all(['p','h1','h2','h3','h4','h5','h6'])
    content = ''
    for p in ps:
        text = p.text
        lines = (line.strip() for line in text.splitlines())
        chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
        text = ' '.join(chunk for chunk in chunks if chunk)
        content += text + '\n'
    # delete - more than 3
    content = re.sub(r'-{3,}', '', content)
    content = re.sub(r'\*{3,}', '', content)
    # remove line with empty content
    content = re.sub(r'^\s+$', '', content, flags=re.MULTILINE)
    # name = soup.find('h1').text.strip()
    matches = re.finditer(r'^(chương [\d\w]+.*|phần thứ [\d\w]+.*)$', content, re.MULTILINE | re.IGNORECASE)
    for i,match in enumerate(matches):
        end = match.end()
        while end < len(content) and content[end] != '\n':
            end += 1
        if end < len(content) and content[end] == '\n':
            content = content[:end] + ':' + content[end+1:]
    return metadata,content,html.prettify()

LAW_IDS=['259700', '454887', '299644', '279974', '357505', '268405', '215628', '333670', '341271', '280020', '489758', '81884', '238506', '411816', '393789', '297783', '443310', '281961', '187083', '142187']

items = []
for url in urls:
    try:
        metadata,text,html = get_text(url)
        print(metadata['ten_van_ban'])
        name_slug = slugify(metadata['ten_van_ban'])[:50]
        items.append({'metadata':metadata,'path':name_slug})
        folder_path = os.path.join('documents', name_slug)
        os.makedirs(folder_path, exist_ok=True)
        with open(os.path.join(folder_path, 'content.txt'), 'w', encoding='utf-8') as f:
            f.write(text)
        with open(os.path.join(folder_path, 'content.html'), 'w', encoding='utf-8') as f:
            f.write(html)
    except Exception as e:
        print(e)
        print(url)
        continue
df = pd.DataFrame(items)
# expose metadata
df = pd.concat([df, df['metadata'].apply(pd.Series)], axis=1)
df['metadata'] = df['metadata'].apply(lambda x: json.dumps(x, ensure_ascii=False))
df.to_csv('documents/data.csv', index=False)

def make_content_for_tree(text):
    re_footer = r'^(Bộ luật này đã được Quốc hội .*|Luật này được Quốc hội.*|Bộ luật này được Quốc hội .*|Luật này đã được Quốc hội .*)'
    matches = re.split(re_footer, text, flags=re.MULTILINE)
    text = matches[0]
    return text

CREATE_CONTENT_TREE=False
df = pd.read_csv('documents/data.csv')
for row in df.itertuples():
    print(row.path)
    folder_path = os.path.join('documents', row.path)
    text = open(os.path.join(folder_path, 'content.txt'), encoding='utf-8').read()
    raw_text = text
    text = make_content_for_tree(text)
    if CREATE_CONTENT_TREE:
        with open(os.path.join(folder_path, 'content_tree.txt'), 'w', encoding='utf-8') as f:
            f.write(text)
    text = open(os.path.join(folder_path, 'content_tree.txt'), encoding='utf-8').read()
    tree = Tree(metadata= json.loads(row.metadata),content=text,raw_text=raw_text)
    with open(os.path.join(folder_path, 'debug_tree.txt'), 'w', encoding='utf-8') as f:
        f.write(str(tree))
    with open(os.path.join(folder_path, 'tree.pkl'), 'wb') as f:
        pickle.dump(tree, f)
    with open(os.path.join(folder_path, 'tree.json'), 'w', encoding='utf-8') as f:
        f.write(json.dumps(tree.export(), indent=4, ensure_ascii=False))
    with open(os.path.join(folder_path, 'tree.json.gz'), 'wb') as f:
        f.write(gzip.compress(json.dumps(tree.export(), ensure_ascii=False).encode('utf-8')))
    # with open('documents/{}.json.gz'.format(row.path), 'wb') as f:
        # f.write(gzip.compress(json.dumps(tree.export(), ensure_ascii=False).encode('utf-8')))
