==> code/2.answers_process.ipynb <==
import string
import pandas as pd
import lawquery
import os

engines = {}
law_df = pd.read_csv('./documents/data.csv')
for i in range(len(law_df)):
    path = law_df['path'][i]
    so_hieu_van_ban = law_df['so_hieu_van_ban'][i]
    if so_hieu_van_ban not in engines:
        engines[so_hieu_van_ban] = lawquery.Engine(
            os.path.join('documents',path, 'tree.json.gz'),
        )
print("Number of engines:",len(engines))

law_df = pd.read_csv('./documents/data.csv')
allow_law_ids = law_df['so_hieu_van_ban'].tolist()
law_df.head()

df = pd.read_json('./answers.jsonl', lines=True)
print(len(df))
df.head()

def answer_convert(s:str):

    if s is None:
        return None
    s = s.strip()
    if s == '':
        return None
    split = s.split('>')
    if len(split) == 1:
        return None
    law_id = split[0].strip()
    if law_id not in allow_law_ids:
        return None
    node_type = split[1].strip().lower()
    if 'điều' not in node_type:
        return None
    node_id = node_type.replace('điều', '').strip()
    # split by . and take first
    qq = engines[law_id].query(node_id=node_id,node_type='điều')
    if len(qq) == 0:
        return None
    node_id = node_id.split('.')[0]
    result={
        'law_id': law_id,
        'node_type':'điều',
        'node_id': node_id,
    }
    return result

df_filtered = []
for (i, row) in df.iterrows():
    answers = row['answers']
    answers = [x.split(' > ')[0].strip() for x in answers]
    allow=False
    for a in answers:
        for allow_law_id in allow_law_ids:
            if allow_law_id in a:
                allow=True
                break
        if allow:   
            break
    if allow:
        new_answers = []
        for a in row['answers']:
            new_answer = answer_convert(a)

            if new_answer is not None:
                try:
                    int(new_answer['node_id'])
                except:
                    print(new_answer)
                new_answers.append(new_answer)
        row['new_answers'] = new_answers
        df_filtered.append(row)
df_filtered = pd.DataFrame(df_filtered)
# drop new_answers is empty
df_filtered = df_filtered[df_filtered['new_answers'].map(len) > 0]
print(len(df_filtered))
df_filtered.head()

# remove duplicate newline
import re

df_filtered['cauhoi'] = df_filtered['cauhoi'].apply(lambda x: re.sub(r'\n+', '\n', x))
df_filtered['cauhoi'] = df_filtered['cauhoi'].apply(lambda x: x.strip())

len(df_filtered)

df_filtered.to_json('./answers_filtered.jsonl', orient='records', lines=True, force_ascii=False)
